lib/db: Remove *instance by making everything *Lowlevel (#6204)
